Backend/requirements.txt
acachecontrol==0.3.7
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
amqp==5.3.1
anyio==4.12.1
asgiref==3.11.0
asttokens==3.0.1
attrs==25.4.0
Authlib==1.6.6
bcrypt==5.0.0
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.4
decorator==5.2.1
Django==5.2.10
django-cors-headers==4.9.0
django-debug-toolbar==6.2.0
django-prometheus==2.4.1
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
executing==2.2.1
factory_boy==3.3.3
Faker==40.1.2
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
kombu==5.6.2
matplotlib-inline==0.2.1
multidict==6.7.1
okta-jwt-verifier==0.3.0
packaging==26.0
parso==0.8.5
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2-binary==2.9.11
pure_eval==0.2.3
pycparser==3.0
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
redis==7.1.0
requests==2.32.5
retry2==0.9.5
six==1.17.0
sqlparse==0.5.5
stack-data==0.6.3
starlette==0.52.1
traitlets==5.14.3
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.5.0
yarl==1.22.0